create new base class for migration class,optimization
