+++ madmigration/main.py
@@ -82,48 +82,50 @@ class Controller:
     def run(self):
         
         # Looping in migrationTables
-        for mt in self.migration_tables:
-            self.destination_table = mt.migrationTable.DestinationTable
+        try:
+            for mt in self.migration_tables:
+                self.destination_table = mt.migrationTable.DestinationTable
 
-            # Migrate class is use based on driver
-            Migrate = detect_driver(self.sourcedb_driver)
+                # Migrate class is use based on driver
+                Migrate = detect_driver(self.sourcedb_driver)
+                
+                # Instance of Migrate class
+                migrate = Migrate(self.config, self.source_db)
+                migrate.parse_migration_tables(mt)    
+
+                # Dictionary is used to keep data about destination column and type_cast format
+                self.convert_info = {}
+
+                # Columns List is used to keep source Columns names
+                # We will send table name and source columns list to function "get_data_from_source_table"
+                # get_data_from_source_table function will yield data with specified columns from row
+                columns = []
+                
+                for column in migrate.columns:
+                    columns.append(column.sourceColumn.get("name"))
+
+                    if column.destinationColumn.options.type_cast:
+                        self.convert_info[
+                            column.destinationColumn.name] = column.destinationColumn.options.type_cast
+                
+                # self.source_data is data received (yield) from get_data_from_source_table function
+                self.source_data = migrate.get_data_from_source_table(mt.migrationTable.SourceTable, columns)
+
+                for source_data in self.source_data:
+                    new_data = Migrate.type_cast(data_from_source=source_data, mt=mt, convert_info=self.convert_info)
+                    Migrate.insert_data(engine=self.destination_db, table_name=self.destination_table.name, data=new_data)
+                
+                # count = 0
+                # from time import sleep
+                # for data in Migrate.loop_in_data(self.source_data, mt, self.convert_info):
+                #     count += 1
+                #     print("count: ",count, "id: ", data.get("id"))
+                #     print(self.destination_table.name)
+                #     sleep(1)
+                    # Migrate.insert_data(engine=self.destination_db, table_name=self.destination_table.name, data=data)
             
-            # Instance of Migrate class
-            migrate = Migrate(mt.migrationTable, self.source_db)
-            migrate.parse_migration_tables()    
-
-            # Dictionary is used to keep data about destination column and type_cast format
-            self.convert_info = {}
-
-            # Columns List is used to keep source Columns names
-            # We will send table name and source columns list to function "get_data_from_source_table"
-            # get_data_from_source_table function will yield data with specified columns from row
-            columns = []
-            
-            for column in migrate.columns:
-                columns.append(column.sourceColumn.get("name"))
-
-                if column.destinationColumn.options.type_cast:
-                    self.convert_info[
-                        column.destinationColumn.name] = column.destinationColumn.options.type_cast
-            
-            # self.source_data is data received (yield) from get_data_from_source_table function
-            self.source_data = migrate.get_data_from_source_table(mt.migrationTable.SourceTable, columns)
-
-            for source_data in self.source_data:
-                new_data = Migrate.type_cast(data_from_source=source_data, mt=mt, convert_info=self.convert_info)
-                Migrate.insert_data(engine=self.destination_db, table_name=self.destination_table.name, data=new_data)
-            
-            # count = 0
-            # from time import sleep
-            # for data in Migrate.loop_in_data(self.source_data, mt, self.convert_info):
-            #     count += 1
-            #     print("count: ",count, "id: ", data.get("id"))
-            #     print(self.destination_table.name)
-            #     sleep(1)
-                # Migrate.insert_data(engine=self.destination_db, table_name=self.destination_table.name, data=data)
-        
-
+        except Exception as err:
+            print(err)
